refactor(channel-management): extract shared lifecycle and polling services
