+++ backend/functions/.shared/channelManagement/channelManagementConstants.js
@@ -1,0 +1,19 @@
+export const SINGLE_ACCOUNT_PER_USER = "SINGLE_ACCOUNT_PER_USER";
+
+export const HOLIDU_STATUS = {
+  NOT_CONNECTED: "NOT_CONNECTED",
+  PENDING_PROVIDER_VALIDATION: "PENDING_PROVIDER_VALIDATION",
+  VALIDATION_FAILED: "VALIDATION_FAILED",
+  RECONNECT_REQUIRED: "RECONNECT_REQUIRED",
+  DISCONNECTED: "DISCONNECTED",
+  CONNECTED: "CONNECTED",
+};
+
+export const CHANNEX_STATUS = {
+  NOT_CONNECTED: "NOT_CONNECTED",
+  PENDING_PROVIDER_VALIDATION: "PENDING_PROVIDER_VALIDATION",
+  VALIDATION_FAILED: "VALIDATION_FAILED",
+  RECONNECT_REQUIRED: "RECONNECT_REQUIRED",
+  DISCONNECTED: "DISCONNECTED",
+  CONNECTED: "CONNECTED",
+};
