+++ backend/functions/.shared/channelManagement/utils/channexBookingPollUtils.js
