+++ backend/functions/.shared/channelManagement/utils/channexBookingRevisionUtils.js
@@ -1,0 +1,5 @@
+export const buildChannexPullIssue = (code, message, extra = {}) => ({
+  code,
+  message,
+  ...extra,
+});
