+++ backend/functions/.shared/integrations/integrationResponse.js
@@ -1,0 +1,8 @@
+export const shapeCredentialIntegrationForResponse = (integration) => {
+  if (!integration || typeof integration !== "object") {
+    return integration;
+  }
+
+  const { credentialsRef, ...safeIntegration } = integration;
+  return safeIntegration;
+};
