+++ backend/functions/UnifiedMessaging/data/messageRepository.js
@@ -1,4 +1,4 @@
-import Database from "../ORM/index.js";
+import Database from "../.shared/integrations/ORM/index.js";
 import { UnifiedMessage } from "database/models/unified/messaging/UnifiedMessage";
 import { randomUUID } from "node:crypto";
